Deferring type resolution to a dedicated `TypeResolverInterface`
