+++ src/ChangeSet/IdentityExtractor/IdentityExtractorInterface.php
@@ -4,9 +4,6 @@ namespace ChangeSet\IdentityExtractor;
 
 interface IdentityExtractorInterface
 {
-    /** @param object $object @return string */
-    public function getType($object);
-
     /** @return mixed|null */
     public function getIdentity($object);
