+++ src/ChangeSet/TypeResolver/TypeResolverInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace ChangeSet\TypeResolver;
+
+interface TypeResolverInterface
+{
+    /**
+     * @param $object
+     *
+     * @return string
+     */
+    public function getTypeOfObject($object);
+
+    /**
+     * @param string $type
+     *
+     * @return string
+     */
+    public function resolveType($type);
+}
